+++ cmd/frontend/internal/httpapi/releasecache/main_test.go
@@ -32,7 +32,7 @@ func TestMain(m *testing.M) {
 func newClientFactory(t testing.TB, name string) (*httpcli.Factory, func(testing.TB)) {
 	cassette := filepath.Join("testdata", strings.ReplaceAll(name, " ", "-"))
 	rec := newRecorder(t, cassette, update(name))
-	mw := httpcli.NewMiddleware(httpcli.GitHubProxyRedirectMiddleware)
+	mw := httpcli.NewMiddleware()
 	return httpcli.NewFactory(mw, httptestutil.NewRecorderOpt(rec)),
 		func(t testing.TB) { save(t, rec) }
 }